feat: portfolio margin, financing, and sizing parity
- portfolio backtests now accept max_leverage, financing rates, and
  every PositionSizing scheme instead of rejecting them, applied at
  the account level (largest-first maintenance liquidation, pro-rata
  margin interest, per-symbol sizing series)
- check_sl_tp is one shared fn, so per-trade brackets fire in
  portfolios and the engines cannot drift again
- engine.rs split into engine/{mod,state,exits,signals,entries,report}
  to hold the 800-line file cap
